Various changes
ContactNumber has been changed from int to long; everything relevant has been adjusted to accommodate the change
DeliveryForm forgot to trim the colons that came with server's new responses;  this has been fixed

+++ src/main/java/com/example/rmonitor/classes/Client.java
@@ -5,9 +5,9 @@ public class Client {
 	String name;
 	String address;
 	String contact_person;
-	int contact_number;
+	long contact_number;
 	
-	public Client(int clientid, String name, String address, String contact_person, int contact_number) {
+	public Client(int clientid, String name, String address, String contact_person, long contact_number) {
 		this.clientid = clientid;
 		this.name = name;
 		this.address = address;
@@ -47,11 +47,11 @@ public class Client {
 		this.contact_person = contact_person;
 	}
 
-	public int getContact_number() {
+	public long getContact_number() {
 		return contact_number;
 	}
 
-	public void setContact_number(int contact_number) {
+	public void setContact_number(long contact_number) {
 		this.contact_number = contact_number;
 	}
 	
@@ -60,7 +60,7 @@ public class Client {
 	}
 
 	public void setContact_number(String contact_number) {
-		this.contact_number = Integer.parseInt(contact_number);
+		this.contact_number = Long.parseLong(contact_number);
 	}
 
 	public boolean isPersisted() {
